+++ Studio/Analysis/ParticleAreaPanel.cpp
@@ -67,9 +67,9 @@ bool ParticleAreaPanel::get_display_particle_area() const {
 }
 
 //---------------------------------------------------------------------------
-Eigen::VectorXf ParticleAreaPanel::get_computed_values() const {
+Eigen::VectorXd ParticleAreaPanel::get_computed_values() const {
   if (!job_) {
-    return Eigen::VectorXf();
+    return Eigen::VectorXd();
   }
   if (ui_->mean_radio->isChecked()) {
     return job_->get_mean_areas();
@@ -158,7 +158,7 @@ void ParticleAreaPanel::update_graphs() {
 
   plot->clearGraphs();
 
-  Eigen::VectorXf numbers;
+  Eigen::VectorXd numbers;
   QString x_label = QString("Mean area");
   QString title = "Mean particle area percent";
   if (ui_->mean_radio->isChecked()) {
