+++ Studio/Analysis/ParticleAreaPanel.h
@@ -38,7 +38,7 @@ class ParticleAreaPanel : public QWidget {
 
   bool get_display_particle_area() const;
 
-  Eigen::VectorXf get_computed_values() const;
+  Eigen::VectorXd get_computed_values() const;
   std::string get_computed_value_name() const;
 
  public Q_SLOTS:
