+++ Studio/Job/GroupPvalueJob.h
@@ -13,12 +13,12 @@ public:
   void run() override;
   QString name() override;
 
-  Eigen::VectorXf get_group_pvalues();
+  Eigen::VectorXd get_group_pvalues();
 
 private:
 
   ParticleShapeStatistics stats_;
-  Eigen::VectorXf group_pvalues_;
+  Eigen::VectorXd group_pvalues_;
 
 };
 }
