+++ Studio/Job/NetworkAnalysisJob.h
@@ -17,14 +17,14 @@ class NetworkAnalysisJob : public Job {
   void set_pvalue_threshold(double pvalue_threshold);
   void set_num_iterations(int num_iterations);
 
-  Eigen::VectorXf get_tvalues();
-  Eigen::VectorXf get_spm_values();
+  Eigen::VectorXd get_tvalues();
+  Eigen::VectorXd get_spm_values();
 
  private:
   std::shared_ptr<Project> project_;
   ParticleShapeStatistics stats_;
-  Eigen::VectorXf tvalues_;
-  Eigen::VectorXf spm_values_;
+  Eigen::VectorXd tvalues_;
+  Eigen::VectorXd spm_values_;
   std::string target_group_;
   std::string target_feature_;
